Refactor: entity 생성 수정
entity 생성 수정
