+++ src/main/java/com/WalkiePaw/domain/board/entity/BoardCategory.java
@@ -1,0 +1,9 @@
+package com.WalkiePaw.domain.board.entity;
+
+/**
+ * JOB_SEARCH = "구직"
+ * JOB_OPENING = "구인"
+ */
+public enum BoardCategory {
+    JOB_SEARCH, JOB_OPENING
+}
