+++ src/main/java/com/WalkiePaw/domain/board/entity/BoardStatus.java
@@ -1,0 +1,10 @@
+package com.WalkiePaw.domain.board.entity;
+
+/**
+ * RECRUITING = "모집 중"
+ * RESERVED = "예약 중"
+ * COMPLETED = "완료"
+ */
+public enum BoardStatus {
+    RECRUITING, RESERVED, COMPLETED
+}
