+++ src/main/java/com/WalkiePaw/domain/chatV1/entity/ChatMessageStatus.java
@@ -1,0 +1,10 @@
+package com.WalkiePaw.domain.chatV1.entity;
+
+
+/**
+ * SEND = "보낸 메시지"
+ * RECEIVE = "받은 메시지"
+ */
+public enum ChatMessageStatus {
+    SEND, RECEIVE
+}
