+++ src/main/java/com/WalkiePaw/domain/chatV2/entity/ChatMessageStatus.java
@@ -1,0 +1,9 @@
+package com.WalkiePaw.domain.chatV2.entity;
+
+/**
+ * SEND = "보낸 메시지"
+ * RECEIVE = "받은 메시지"
+ */
+public enum ChatMessageStatus {
+    SEND, RECEIVE
+}
