+++ src/main/java/com/WalkiePaw/domain/chatroom/entity/ChatroomStatus.java
@@ -1,0 +1,10 @@
+package com.WalkiePaw.domain.chatroom.entity;
+
+/**
+ * GENERAL = "일반"
+ * RESERVED = "예약"
+ * COMPLETED = "거래완료"
+ */
+public enum ChatroomStatus {
+    GENERAL, RESERVED, COMPLETED
+}
