--- src/main/java/com/WalkiePaw/domain/entity/BaseEntity.java
@@ -1,22 +1,0 @@
-package com.WalkiePaw.domain.entity;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.EntityListeners;
-import jakarta.persistence.MappedSuperclass;
-import lombok.Getter;
-import org.springframework.data.annotation.CreatedDate;
-import org.springframework.data.annotation.LastModifiedDate;
-import org.springframework.data.jpa.domain.support.AuditingEntityListener;
-
-import java.time.LocalDate;
-
-@MappedSuperclass
-@EntityListeners(AuditingEntityListener.class)
-@Getter
-public abstract class BaseEntity {
-    @CreatedDate
-    @Column(updatable = false)
-    private LocalDate createdDate;
-    @LastModifiedDate
-    private LocalDate modifiedDate;
-}
