--- src/main/java/com/WalkiePaw/domain/entity/board/BoardCategory.java
@@ -1,9 +1,0 @@
-package com.WalkiePaw.domain.entity.board;
-
-/**
- * JOB_SEARCH = "구직"
- * JOB_OPENING = "구인"
- */
-public enum BoardCategory {
-    JOB_SEARCH, JOB_OPENING
-}
