--- src/main/java/com/WalkiePaw/domain/entity/board/BoardStatus.java
@@ -1,10 +1,0 @@
-package com.WalkiePaw.domain.entity.board;
-
-/**
- * RECRUITING = "모집 중"
- * RESERVED = "예약 중"
- * COMPLETED = "완료"
- */
-public enum BoardStatus {
-    RECRUITING, RESERVED, COMPLETED
-}
