--- src/main/java/com/WalkiePaw/domain/entity/chatV1/ChatMessageStatus.java
@@ -1,10 +1,0 @@
-package com.WalkiePaw.domain.entity.chatV1;
-
-
-/**
- * SEND = "보낸 메시지"
- * RECEIVE = "받은 메시지"
- */
-public enum ChatMessageStatus {
-    SEND, RECEIVE
-}
