--- src/main/java/com/WalkiePaw/domain/entity/chatV2/ChatMessageStatus.java
@@ -1,9 +1,0 @@
-package com.WalkiePaw.domain.entity.chatV2;
-
-/**
- * SEND = "보낸 메시지"
- * RECEIVE = "받은 메시지"
- */
-public enum ChatMessageStatus {
-    SEND, RECEIVE
-}
