--- src/main/java/com/WalkiePaw/domain/entity/chatroom/ChatroomStatus.java
@@ -1,10 +1,0 @@
-package com.WalkiePaw.domain.entity.chatroom;
-
-/**
- * GENERAL = "일반"
- * RESERVED = "예약"
- * COMPLETED = "거래완료"
- */
-public enum ChatroomStatus {
-    GENERAL, RESERVED, COMPLETED
-}
