--- src/main/java/com/WalkiePaw/domain/entity/qna/QnaStatus.java
@@ -1,10 +1,0 @@
-package com.WalkiePaw.domain.entity.qna;
-
-/**
- * WAITING = "답변 대기 중"
- * COMPLETED = "답변 완료"
- */
-
-public enum QnaStatus {
-    WAITING, COMPLETED
-}
