+++ src/main/java/com/WalkiePaw/domain/member/MemberRepository.java
@@ -1,0 +1,31 @@
+package com.WalkiePaw.domain.member;
+
+import com.WalkiePaw.domain.member.entity.Member;
+import jakarta.persistence.EntityManager;
+import jakarta.persistence.PersistenceContext;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Repository;
+
+import java.util.List;
+
+@Repository
+@RequiredArgsConstructor
+public class MemberRepository {
+
+    @PersistenceContext
+    private final EntityManager em;
+
+    public void save(Member member) {
+        em.persist(member);
+    }
+
+    public Member findOne(Integer memberId) {
+        return em.find(Member.class, memberId);
+    }
+
+    public List<Member> findByName(String name) {
+        return em.createQuery("select m from Member m where m.name = :name", Member.class)
+                .setParameter("name", name)
+                .getResultList();
+    }
+}
