+++ src/main/java/com/WalkiePaw/domain/member/entity/MemberStatus.java
@@ -1,0 +1,12 @@
+package com.WalkiePaw.domain.member.entity;
+
+/**
+ * GENERAL = "일반"
+ * WITHDRAWN = "탈퇴"
+ * BANNED = "제재 또는 정지된 회원"
+ * INACTIVE = "휴면 상태"
+ */
+public enum MemberStatus {
+    GENERAL, WITHDRAWN, BANNED, INACTIVE
+
+}
