+++ src/main/java/com/WalkiePaw/domain/qna/entity/QnaStatus.java
@@ -1,0 +1,10 @@
+package com.WalkiePaw.domain.qna.entity;
+
+/**
+ * WAITING = "답변 대기 중"
+ * COMPLETED = "답변 완료"
+ */
+
+public enum QnaStatus {
+    WAITING, COMPLETED
+}
