+++ src/main/java/com/WalkiePaw/domain/report/entity/BoardReportCategory.java
@@ -1,0 +1,13 @@
+package com.WalkiePaw.domain.report.entity;
+
+/**
+ * SPAM = 스팸홍보/도배글
+ * PORNOGRAPHY = 음란물
+ * ILLEGAL_CONTENT = 불법정보를 포함
+ * NOXIOUS = 청소년에게 유해한 내용
+ * HARASSMENT = 불쾌한 표현
+ * ETC = 기타
+ */
+public enum BoardReportCategory {
+    SPAM, PORNOGRAPHY, ILLEGAL_CONTENT, NOXIOUS, HARASSMENT, ETC
+}
